refactor: restructuring folders
for a more 'feature-oriented architecture'
