+++ src/app/game/game-multiplayer/multiplayer.component.ts
@@ -2,8 +2,8 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { trigger, style, animate, transition } from '@angular/animations';
 import { Router } from '@angular/router';
 
-import { MultiplayerService, GAMESTATE } from './services/multiplayer.service';
-import { WebSocketService } from './services/web-socket.service';
+import { MultiplayerService, GAMESTATE } from '../services/multiplayer.service';
+import { WebSocketService } from '../services/web-socket.service';
 import { routes } from '../../shared/models/routes';
 import { Subscription } from 'rxjs';
 import { LobbyComponent } from './lobby/lobby.component';
