+++ src/app/game/game-pick-difficulty/game-pick-difficulty.component.ts
@@ -1,15 +1,15 @@
 import { Component, Output, EventEmitter } from '@angular/core';
 import { Router } from '@angular/router';
 import { routes } from '../../shared/models/routes';
-import { MultiplayerService, GAMESTATE } from '../game-multiplayer/services/multiplayer.service';
-import { WebSocketService } from '../game-multiplayer/services/web-socket.service';
+import { MultiplayerService, GAMESTATE } from '../services/multiplayer.service';
+import { WebSocketService } from '../services/web-socket.service';
 import { SocketEndpoints } from '../../shared/models/websocketEndpoints';
 import { MatButton } from '@angular/material/button';
 import { MatIcon } from '@angular/material/icon';
 import { NgIf } from '@angular/common';
-import { GameLevelConfig, GameConfigService } from '../game-config.service';
-import { TranslationService } from '@/app/services/translation.service';
-import { TranslatePipe } from '@/app/pipes/translation.pipe';
+import { GameLevelConfig, GameConfigService } from '../services/game-config.service';
+import { TranslationService } from '@/app/core/translation.service';
+import { TranslatePipe } from '@/app/core/translation.pipe';
 
 @Component({
     selector: 'app-game-pick-difficulty',
