+++ src/app/game/game-word-to-draw/game-word-to-draw.component.ts
@@ -1,17 +1,17 @@
 import { Component, OnInit, Output, EventEmitter, OnDestroy } from '@angular/core';
 import { interval, Subscription } from 'rxjs';
 import { take, tap } from 'rxjs/operators';
-import { DrawingService } from '../game-draw/services/drawing.service';
+import { DrawingService } from '../services/drawing.service';
 import { Router } from '@angular/router';
 import { routes } from '../../shared/models/routes';
-import { MultiplayerService, GAMESTATE } from '../game-multiplayer/services/multiplayer.service';
+import { MultiplayerService, GAMESTATE } from '../services/multiplayer.service';
 import { MatIcon } from '@angular/material/icon';
 import { MatButton } from '@angular/material/button';
 import { MatProgressSpinner } from '@angular/material/progress-spinner';
 import { UpperCasePipe } from '@angular/common';
-import { GameConfigService } from '../game-config.service';
-import { TranslationService } from '@/app/services/translation.service';
-import { TranslatePipe } from '@/app/pipes/translation.pipe';
+import { GameConfigService } from '../services/game-config.service';
+import { TranslationService } from '@/app/core/translation.service';
+import { TranslatePipe } from '@/app/core/translation.pipe';
 
 @Component({
   selector: 'app-game-word-to-draw',
@@ -27,7 +27,7 @@ import { TranslatePipe } from '@/app/pipes/translation.pipe';
   ],
 })
 export class GameWordToDrawComponent implements OnInit, OnDestroy {
-  config = this.gameConfigService.getConfig; //getting the current/default set game level values (easy mode)
+  config = this.gameConfigService.getConfig; 
   
   isSinglePlayer = false;
   isMultiPlayer = false;
