+++ src/app/game/idle-timeout/idle-timeout.component.html
