+++ src/app/game/services/image.service.ts
