+++ src/main.ts
@@ -9,7 +9,7 @@ import { bootstrapApplication } from '@angular/platform-browser';
 import { InfoDialogComponent } from './app/admin/info-dialog/info-dialog.component';
 import { withInterceptorsFromDi, provideHttpClient, HttpClient } from '@angular/common/http';
 import { provideRouter } from '@angular/router';
-import { TranslationService } from './app/services/translation.service';
+import { TranslationService } from './app/core/translation.service';
 
 if (environment.production) {
   enableProdMode();
